requirements.txt
absl-py==1.4.0
astunparse==1.6.3
blessed==1.20.0
blinker==1.8.2
cached-property==1.5.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.1.0
clang==16.0.1.1
click==8.1.3
cmake==3.26.4
contourpy==1.0.7
cycler==0.11.0
Cython==3.0.8
dataclasses==0.6
easydict==1.11
filelock==3.12.0
Flask==3.0.3
flatbuffers==23.5.26
fonttools==4.39.4
fsspec==2023.5.0
ftfy==6.1.1
gast==0.4.0
google-auth==2.19.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gpustat==1.1
grad-cam==1.3.7
grpcio==1.54.2
h5py==3.8.0
huggingface-hub==0.14.1
idna==3.4
imageio==2.33.1
imgaug==0.4.0
importlib-metadata==6.6.0
importlib-resources==5.12.0
itsdangerous==2.2.0
jax==0.4.10
Jinja2==3.1.2
joblib==1.2.0
keras==2.12.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lazy_loader==0.3
libclang==16.0.0
lit==16.0.6
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
ml-dtypes==0.1.0
mpmath==1.3.0
msgpack==1.0.7
networkx==3.1
nltk==3.8.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==11.525.112
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
nvitop==1.1.2
oauthlib==3.2.2
olefile==0.46
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
protobuf==4.23.2
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-crf==0.7.2
pytorch-gradcam==0.2.1
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sacremoses==0.0.53
safetensors==0.3.1
scikit-image==0.22.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
sentencepiece==0.1.99
senticnet==1.6
shapely==2.0.2
six==1.16.0
sympy==1.12
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.12.9
timm @ file:///home/wdq/src/pytorch-image-models
tokenizers==0.13.3
torch==2.1.0
torchaudio==2.0.1
torchdata==0.7.0
torchtext==0.16.0
torchvision==0.15.2+cu118
tqdm==4.65.0
transformers==4.29.2
triton==2.1.0
ttach==0.0.3
typing_extensions==4.6.2
tzdata==2023.3
urllib3==1.26.16
wcwidth==0.2.6
Werkzeug==3.0.3
wrapt==1.14.1
zipp==3.15.0
